Split Replica.VDOM module
